[CPBugFix-002][UpdatedDocs] Fixed the bug, updated some docs, and re-organized a bit.
