+++ .swiftpm/xcode/xcshareddata/xcschemes/commitPrefix.xcscheme
@@ -52,6 +52,16 @@
                ReferencedContainer = "container:">
             </BuildableReference>
          </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "CommitPrefixTests"
+               BuildableName = "CommitPrefixTests"
+               BlueprintName = "CommitPrefixTests"
+               ReferencedContainer = "container:">
+            </BuildableReference>
+         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
