+++ Sources/CommitPrefix/Hook/CommitMessageHook.swift
@@ -1,0 +1,117 @@
+//
+//  CommitMessageHook.swift
+//  commitPrefix
+//
+//  MIT License
+//
+//  Copyright (c) 2019 STEPHEN L. MARTINEZ
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in all
+//  copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//  SOFTWARE.
+
+import Files
+import Foundation
+
+struct CommitMessageHook {
+    
+    private let hooksDirectory: Folder
+    private let cmHookContents: CommitMessageHookContents
+    
+    private init(gitDirectory: Folder) throws {
+        guard let hooksDirectory = try? gitDirectory.subfolder(named: FolderName.hooks) else {
+            throw CPError.directoryNotFound(name: FolderName.hooks, path: gitDirectory.path)
+        }
+        self.hooksDirectory = hooksDirectory
+        self.cmHookContents = CommitMessageHookContents()
+    }
+    
+    /// A commit-message hook is required for this application to work properly. This method
+    /// checks to see if the hook can be located, if it is one that is generated by this
+    /// application and then generates one if it can't be found.
+    ///
+    /// - parameters:
+    ///     - gitDirectory: A  `Folder` representing the git directory
+    ///
+    static func findOrCreate(with gitDirectory: Folder) throws {
+        let cpHook = try CommitMessageHook(gitDirectory: gitDirectory)
+        try cpHook.locateOrCreateHook()
+    }
+    
+    private func getCommitHookFile() throws -> File? {
+        
+        guard let foundCommitHookFile = try? hooksDirectory.file(named: FileName.commitMessage) else {
+        
+            do {
+                let commitHookFile = try hooksDirectory.createFile(named: FileName.commitMessage)
+                try commitHookFile.write(cmHookContents.renderScript(), encoding: .utf8)
+                cpDebugPrint(commitHookFile.path)
+                Shell.makeExecutable(commitHookFile.path)
+            } catch {
+                throw CPError.hookReadWriteError
+            }
+            
+            return nil
+            
+        }
+        
+        return foundCommitHookFile
+        
+    }
+    
+    private func overwriteCommitHook(_ commitHookFile: File) throws {
+        print("There seems to be an existing commit-msg found in the hooks directory")
+        print("Would you like to overwrite? [y/n]")
+        let answer = readLine() ?? ""
+        
+        switch answer {
+            
+        case "y":
+            print("Overwritting existing commit-msg with generated hook")
+            
+            do {
+                // TODO: - Theres a case where the file is not executable in the first place this will not correct that
+                try commitHookFile.write(cmHookContents.renderScript(), encoding: .utf8)
+            } catch {
+                throw CPError.hookReadWriteError
+            }
+            
+        case "n":
+            throw CPTermination.overwriteCancelled
+            
+        default:
+            throw CPTermination.expectedYesOrNo
+            
+        }
+    }
+    
+    private func hookIsCommitPrefix(_ hookFile: File) throws -> Bool {
+        
+        guard let hookContents = try? hookFile.readAsString(encodedAs: .utf8) else {
+            throw CPError.hookReadWriteError
+        }
+        
+        return hookContents.contains(cmHookContents.fileIdentifier)
+    }
+    
+    private func locateOrCreateHook() throws {
+        guard let foundCommitHookFile = try getCommitHookFile() else { return }
+        guard try !hookIsCommitPrefix(foundCommitHookFile) else { return }
+        try overwriteCommitHook(foundCommitHookFile)
+    }
+    
+}
